@@ -1,20 +1,20 @@
 name: copy_with_e_generator
 description: Creates copyWith extension methods for each annotated class and works with abstract too.
-version: 1.0.10-nullsafety5
+version: 1.0.10-nullsafety13
 homepage: https://github.com/atreeon/CopyWithE
 
 environment:  
-  sdk: ">=2.7.2 <3.0.0"
+  sdk: ">=2.12.0-0 <3.0.0"
 
 dependencies:
   analyzer_models: ^1.0.3
-  build: ^1.2.2
-  source_gen: ^0.9.4+7
+  build: ^1.6.3
+  source_gen: ^0.9.10+3
   copy_with_e_annotation: ^1.0.2-nullsafety2
-  dartx: '>=0.3.0 <0.7.0'
+  dartx: ^0.6.0
 #  tuple: ^1.0.3
-  basic_utils: ^2.4.8
-  analyzer: '>=0.40.0 <0.42.0'
+  analyzer: '>=0.40.0 <=1.0.0'
+  adi_helpers: ^1.0.7-nullsafety2
 
 dev_dependencies:  
   build_runner: ^1.7.3
