@@ -4,8 +4,8 @@ import 'package:copy_with_e_annotation/copy_with_e_annotation.dart';
 
 @CopyWithE()
 class Person {
-  final int age;
-  final String name;
+  final int? age;
+  final String? name;
 
   Person({this.age, this.name});
 
